verilator: Add trace and print support to verilator bench
The verilator bench did not support instruction tracing as
it does not include the verilog monitor code used with the icarus bench.
This patch adds tracing by bringing in some code from or1ksim to handle
the tracing.

Also, we enable interpreting the nop putc monitor hook to allow printing
output in verilator even when UART is disabled.